+++ src/gvt/render/data/Dataset.cpp
@@ -70,12 +70,12 @@ bool Dataset::intersect(Ray &r, isecDomList &inter) {
   return false;
 }
 
-AbstractDomain *Dataset::getDomain(int id) {
+AbstractDomain *Dataset::getDomain(size_t id) {
   GVT_ASSERT_BACKTRACE(id < domainSet.size(), "Getting domain outside bound");
   return domainSet[id];
 }
 
-Light *Dataset::getLight(int id) {
+Light *Dataset::getLight(size_t id) {
   GVT_ASSERT_BACKTRACE(id < lightSet.size(),
                        "Getting light source outside bound");
   return lightSet[id];
